# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SIFT - Personal Spending Behavioral Analysis API
  version: 1.0.0
  description: |
    API-first backend for the SIFT spending behavior analysis app.
    All frontend clients (web, future iOS) consume these endpoints.
    Deterministic computation precedes all AI usage.
    LLM cannot generate or modify numeric outputs.
servers:
  - url: /v1
    description: V1 API
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    Account:
      type: object
      required: [id, name, type, source]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        official_name:
          type: string
        type:
          type: string
        subtype:
          type: string
        mask:
          type: string
        current_balance:
          type: number
        available_balance:
          type: number
        iso_currency_code:
          type: string
        source:
          type: string
          enum: [plaid, csv, pdf]
...
    PlaidLinkToken:
      type: object
      required: [link_token, expiration]
      properties:
        link_token:
          type: string
        expiration:
          type: string
          format: date-time
...
paths:
...
  /plaid/link-token:
    post:
      tags: [Plaid]
      summary: Create Plaid link token for frontend
      responses:
        '200':
          description: Link token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PlaidLinkToken'
              example:
                data:
                  link_token: "link-sandbox-abc123"
                  expiration: "2026-02-24T04:00:00Z"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plaid/exchange:
    post:
      tags: [Plaid]
      summary: Exchange public token and sync transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_token]
              properties:
                public_token:
                  type: string
            example:
              public_token: "public-sandbox-abc123"
      responses:
        '200':
          description: Token exchanged, sync started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      connection_id:
                        type: string
                      accounts:
                        type: array
                        items:
                          $ref: '#/components/schemas/Account'
                      import_job_id:
                        type: string
              example:
                data:
                  connection_id: "550e8400-e29b-41d4-a716-446655440000"
                  accounts:
                    - id: "660e8400-e29b-41d4-a716-446655440001"
                      name: "Checking"
                      type: "depository"
                      subtype: "checking"
                      source: "plaid"
                  import_job_id: "770e8400-e29b-41d4-a716-446655440002"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plaid/webhook:
    post:
      tags: [Plaid]
      summary: Plaid webhook endpoint
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
...